requirements.txt
numpy>=1.24.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.30.0
opencv-python>=4.8.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0
tqdm>=4.65.0
scipy>=1.11.0
matplotlib>=3.7.0
Pillow>=10.0.0